fix: refactored lib into command and github
Split code out of generic module to keep things organized.
